requirements.txt
requests~=2.32.3
beautifulsoup4~=4.12.3
EbookLib~=0.18
pillow~=10.4.0
flask~=3.0.2
werkzeug~=3.0.1
eyed3~=0.9.7
PyJWT~=2.8.0
paramiko~=3.4.0
ffmpeg~=1.4
moviepy~=1.0.3
mutagen~=1.47.0
wakeonlan~=3.1.0
psutil~=5.9.8
yt-dlp>=2024.7.7
sqlalchemy~=2.0.35
flask-sqlalchemy~=3.1.1
alembic~=1.13.3